comment unused codes and fix some warnings
