+++ package.json
@@ -23,6 +23,7 @@
     "material-ui-dropzone": "^3.5.0",
     "merge-images": "^2.0.0",
     "node-forge": "1.2.1",
+    "nth-check": "2.0.1",
     "os-browserify": "^0.3.0",
     "path-browserify": "^1.0.1",
     "randomstring": "^1.1.5",
