+++ src/components/App/Cases/PostCase/Finalisation/Finalisation.js
@@ -11,7 +11,7 @@ import SaveIcon from '@material-ui/icons/Save'
 import Grid from '@material-ui/core/Grid'
 import {makeStyles} from '@material-ui/core/styles'
 import TextField from '@material-ui/core/TextField'
-import MenuItem from '@material-ui/core/MenuItem'
+// import MenuItem from '@material-ui/core/MenuItem'
 import oStyle from '../../../../UI/ResponsiveDesign/AuthStyle'
 import {UPDATE_LEVEL, START_LOADER, STOP_LOADER, UPDATE_STEPPER_POSTCASE} from '../../../../../store/actions'
 import config from '../../../../../config'
