+++ src/components/UI/ResponsiveDesign/ResponsiveCard.js
@@ -1,8 +1,7 @@
-import React from 'react'
 import { withStyles } from '@material-ui/core/styles'
 import Card from '@material-ui/core/Card'
 
-import ResponsiveConstants from './ResponsiveConstants'
+import {ResponsiveConstants} from './ResponsiveConstants'
 
 const styleSheet = theme => ({
   root: {
